Moved AbstractEntity to entities package
By grouping all entity related stuff in one place, we can make the
project easier to read for someone who just generated that.

That said, AbstractEntity is something that we should not need to
generate to the project. Can we find just utility class ready made for
us? Would Spring provide us with something that would fill this gap?