Removed CrudEndpoint to simplify generated project
To keep generated project simple to learn and understand, we should
stick to basics. When Enpoints are inherited, we fail to teach how
simple endpoints can be built. If user later on wants to add abstract
parents to re-use functionality, we should explain that in documentation
teaching about EndpointExposed there.